requirements.txt
streamlit
pillow
numpy
matplotlib
ultralytics
opencv-contrib-python-headless
python-dotenv
google-generativeai
